requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
jupyter>=1.0.0
pytest>=7.0.0
